split out procoess schema defs
